Clippy fixes

+++ objdiff-wasm/src/api.rs
@@ -225,7 +225,7 @@ impl GuestDisplay for Component {
         let obj_diff = diff.get::<ResourceObjectDiff>();
         let obj = obj_diff.0.as_ref();
         let symbol_display = from_symbol_ref(symbol_ref);
-        diff::display::symbol_context(obj, symbol_display.symbol as usize)
+        diff::display::symbol_context(obj, symbol_display.symbol)
             .into_iter()
             .map(ContextItem::from)
             .collect()
@@ -237,7 +237,7 @@ impl GuestDisplay for Component {
         let addend = 0; // TODO
         let override_color = None; // TODO: colorize replaced/deleted/inserted relocations
         let symbol_display = from_symbol_ref(symbol_ref);
-        diff::display::symbol_hover(obj, symbol_display.symbol as usize, addend, override_color)
+        diff::display::symbol_hover(obj, symbol_display.symbol, addend, override_color)
             .into_iter()
             .map(HoverItem::from)
             .collect()
